cache_volume, controller: move pvc helpers to its own file
