+++ pkg/edgecache/cache_volume/controller.go
@@ -18,7 +18,6 @@ package cachevolume
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"time"
 
@@ -34,7 +33,6 @@ import (
 	corelisters "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
-	"k8s.io/client-go/util/retry"
 	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/blob-csi-driver/pkg/edgecache"
@@ -42,115 +40,15 @@ import (
 )
 
 const (
-	finalizerName          string = "external/blob-csi-driver-edgecache"
-	pvcProtectionFinalizer string = "kubernetes.io/pvc-protection"
-	accountAnnotation      string = "external/edgecache-account"
-	containerAnnotation    string = "external/edgecache-container"
-    createVolumeAnnotation string = "external/edgecache-create-volume"
-    secretNameAnnotation   string = "external/edgecache-secret-name"
-    secretKeyField         string = "azurestorageaccountkey"
+	finalizerName             string = "external/blob-csi-driver-edgecache"
+	pvcProtectionFinalizer    string = "kubernetes.io/pvc-protection"
+	accountAnnotation         string = "external/edgecache-account"
+	containerAnnotation       string = "external/edgecache-container"
+	createVolumeAnnotation    string = "external/edgecache-create-volume"
+	secretNameAnnotation      string = "external/edgecache-secret-name"
+	secretNamespaceAnnotation string = "external/edgecache-secret-namespace"
+	secretKeyField            string = "azurestorageaccountkey"
 )
-
-// PVCModification is a type of function that modifies a PVC
-type PVCModification func(*v1.PersistentVolumeClaim) *v1.PersistentVolumeClaim
-
-// PVModification is a type of function that modifies a PV
-type PVModification func(*v1.PersistentVolume) *v1.PersistentVolume
-
-func GetPVByVolumeID(client clientset.Interface, volumeID string) (*v1.PersistentVolume, error) {
-	klog.V(3).Infof("No pvName provided, looking up via volumeID: %s", volumeID)
-	pvList, err := client.CoreV1().PersistentVolumes().List(context.TODO(), metav1.ListOptions{})
-	if err != nil {
-		klog.Errorf("unable to list volumes via volumeID: %s", volumeID)
-		return nil, err
-	}
-	for _, pv := range pvList.Items {
-		if pv.Spec.CSI.VolumeHandle == volumeID {
-			return &pv, nil
-		}
-	}
-	return nil, errors.New("no pv found")
-}
-
-func GetPVByName(client clientset.Interface, pvName string) (*v1.PersistentVolume, error) {
-	pv, err := client.CoreV1().PersistentVolumes().Get(context.TODO(), pvName, metav1.GetOptions{})
-	if err != nil {
-		klog.Errorf("unable to get PV %s", pvName)
-		return nil, err
-	}
-	return pv, nil
-}
-
-func GetPVCByName(client clientset.Interface, namespace string, pvcName string) (*v1.PersistentVolumeClaim, error) {
-	pvc, err := client.CoreV1().PersistentVolumeClaims(namespace).Get(context.TODO(), pvcName, metav1.GetOptions{})
-	if err != nil {
-		klog.Errorf("unable to get PVC %s", pvcName)
-		return nil, err
-	}
-	return pvc, nil
-}
-
-// Attempts to update a pvc, retries if pvc isn't fresh
-func RetryUpdatePVC(client clientset.Interface, namespace string, pvcName string, fn PVCModification) error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		// Fetch the resource here; you need to refetch it on every try, since
-		// if you got a conflict on the last update attempt then you need to get
-		// the current version before making your own changes.
-		pvc, err := client.CoreV1().PersistentVolumeClaims(namespace).Get(context.TODO(), pvcName, metav1.GetOptions{})
-		if err != nil {
-			klog.Errorf("unable to get PVC %s", pvcName)
-			return err
-		}
-
-		// Make whatever updates to the resource are needed
-		pvc = fn(pvc)
-
-		// Try to update
-		_, err = client.CoreV1().PersistentVolumeClaims(namespace).Update(context.TODO(), pvc, metav1.UpdateOptions{})
-		// You have to return err itself here (not wrapped inside another error)
-		// so that RetryOnConflict can identify it correctly.
-		return err
-	})
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		klog.Errorf("RetryOnConflict failed with: %s", err)
-		return err
-	}
-
-	return nil
-}
-
-// Attempts to update a pv, retries if pv isn't fresh
-func RetryUpdatePV(client clientset.Interface, pvName string, fn PVModification) error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		// Fetch the resource here; you need to refetch it on every try, since
-		// if you got a conflict on the last update attempt then you need to get
-		// the current version before making your own changes.
-		pv, err := client.CoreV1().PersistentVolumes().Get(context.TODO(), pvName, metav1.GetOptions{})
-		if err != nil {
-			klog.Errorf("unable to get PV %s", pvName)
-			return err
-		}
-
-		// Make whatever updates to the resource are needed
-		pv = fn(pv)
-
-		// Try to update
-		_, err = client.CoreV1().PersistentVolumes().Update(context.TODO(), pv, metav1.UpdateOptions{})
-		// You have to return err itself here (not wrapped inside another error)
-		// so that RetryOnConflict can identify it correctly.
-		return err
-	})
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		klog.Errorf("RetryOnConflict failed with: %s", err)
-		return err
-	}
-
-	return nil
-}
 
 // Controller is controller that removes PVProtectionFinalizer
 // from PVs that are not bound to PVCs.
